Add files via upload
Screenshots of the results of my VBA program on the three years in the spreadsheet, 1 through 3. VBA_Challenge.vb is the Visual Basic script I had saved in Visual Studio Code for editing - I also have a .bas file if that's easier for you, just let me know and I'll submit that as well if need be.